requirements.txt
alembic==1.7.7
amqp==5.1.1
anyio==3.6.1
apache-airflow==2.3.1
apache-airflow-providers-celery==2.1.4
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==20.3.0
Babel==2.10.1
backcall==0.2.0
billiard==3.6.4.0
blinker==1.4
cachelib==0.7.0
cattrs==1.10.0
celery==5.2.6
certifi==2020.12.5
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
configparser==5.3.0
ConfigUpdater==3.1.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==36.0.2
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.2.1
docutils==0.18.1
email-validator==1.2.1
entrypoints==0.4
exceptiongroup==1.0.1
executing==1.1.1
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flower==1.0.0
git-filter-repo==2.38.0
graphviz==0.20
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
humanize==4.1.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
inflection==0.5.1
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==1.1.0
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.5.1
jupyter-client==7.4.2
jupyter-core==4.11.1
kombu==5.2.4
lazy-object-proxy==1.7.1
linkify-it-py==2.0.0
lockfile==0.12.2
Mako==1.2.0
Markdown==3.3.7
markdown-it-py==2.1.0
MarkupSafe==2.0.1
marshmallow==3.15.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.1
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.23.3
oauthlib==3.2.1
packaging==21.3
pandas==1.5.0
parso==0.8.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==24.0.1
redis==4.3.4
requests==2.27.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==12.6.0
setproctitle==1.2.3
six==1.16.0
sniffio==1.2.0
spotipy==2.20.0
SQLAlchemy==1.4.16
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.3
stack-data==0.5.1
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
tornado==6.1
traitlets==5.4.0
typing-extensions==4.2.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.14.1
WTForms==2.3.3
zipp==3.8.0
